refactor(urls): separate types from config values
Improve architecture by separating concerns in @duyet/urls package:

## Structure

- `types.ts`: All TypeScript type definitions
  - AppUrls, ExternalUrls, UrlsConfig
  - NavLink, Navigation
  - DeepPartial utility type

- `duyet.urls.ts`: Duyet's URL configuration values
  - Environment-aware app URLs
  - External project URLs
  - Clean separation of data from types

- `utils.ts`: Helper functions
  - createUrls(): URL configuration merging
  - createNavigation(): Navigation generation
  - getAppUrls(), getAppUrl(): URL accessors

- `index.ts`: Clean export barrel
  - Re-exports all types and utilities
  - Central entry point

## Benefits

- Clear separation of concerns
- Types separate from values
- Easier to understand and maintain
- Matches profile package pattern
- Better IDE autocomplete

## Usage

Same as before, fully backward compatible:

```ts
import { duyetUrls, createNavigation } from '@duyet/urls'
import type { UrlsConfig } from '@duyet/urls'

// Or import from specific files:
import { duyetUrls } from '@duyet/urls/duyet'
import type { UrlsConfig } from '@duyet/urls/types'
```
