+++ packages/urls/package.json
@@ -6,7 +6,10 @@
   "main": "./src/index.ts",
   "types": "./src/index.ts",
   "exports": {
-    ".": "./src/index.ts"
+    ".": "./src/index.ts",
+    "./types": "./src/types.ts",
+    "./duyet": "./src/duyet.urls.ts",
+    "./utils": "./src/utils.ts"
   },
   "scripts": {
     "lint": "eslint .",
